Don't instantiate globals in bp_groupblog_member_join()
bp_groupblog_member_join() erroneously called a number of globals, including
$blog_id, which it then proceeded to overwrite with local data. This resulted
in some very strange behavior when add_user_to_blog() was run later on,
because switch_to_blog() (which relies on the $blog_id global) got confused
about where it should switch from/to. I have a feeling that this is at the
root of a lot of groupblog issues we've had in the past.