Improve the logic around adding users to sites on group status modification.
By ensuring the default roles are always set, we avoid issues where an
empty role is passed to `add_user_to_blog()`, resulting in the user
being assigned the Subscriber role.

In addition, there's no reason to run the "promote" code when the
'silent_add' option is not checked. This is likely the cause of many
bugs around the internet.